Changed the src/topics/create.js because of incorrect changes as a result of the create.ts being in typescript, thus the file other than changes was reverted to js
